fix: modify cargo
